Update wifi.sh
fixed regex. remember {0,32} only applies to the string NOT to the variable so it only limits cases caught of the regex selected. So in this case it was only selected the last 32 occurrences in the string wifinetworks that was greater than 32 characters. Should use input sanitization to check for character limits. 

Going to do final test to check if this is working with new branch changes and then approve it for master.